changed create statements to include auto gen for the primary key.

+++ sql/ddl/create_respondent_type.sql
@@ -1,7 +1,7 @@
 DROP TABLE IF EXISTS dev_catalog.slv_cdm_hrs.respondent_type;
 
 CREATE TABLE dev_catalog.slv_cdm_hrs.respondent_type (
-    respondent_type_id INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    respondent_type_id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     respondent_type_description STRING,
     start_date DATE,
     end_date DATE,
